requirements.txt
requests==2.31.0
Flask==3.0.2
feedparser==6.0.11
html2text==2024.2.26
beautifulsoup4==4.12.3
openai==0.28
nltk==3.8.1
transformers==4.39.1
opencc-python-reimplemented==0.1.7
dirtyjson==1.0.8
python-dotenv==1.0.1
sentencepiece==0.2.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
lxml==5.1.0
sacremoses==0.1.1
python-telegram-bot==21.0.1
pymongo[srv]
